refactor!: many changes to utils:
* move grid stuff to this own file
* cache calls to cosmo funcs
* make the grid form skymap function generic for any wcs header
