+++ witch/utils.py
@@ -1,10 +1,9 @@
 """
 A set of utility functions and constants used for unit conversions
-and adding generic structure common to multiple models.
+and cosmology as well as some generically useful math functions.
 """
 
-from functools import partial
-from typing import Optional
+from functools import cache, partial
 
 import jax
 import jax.numpy as jnp
@@ -13,46 +12,32 @@ from astropy import constants as const
 from astropy import units as u
 from astropy.cosmology import Planck15 as cosmo
 from jax.typing import ArrayLike
-from numpy.typing import NDArray
 
-Grid = tuple[jax.Array, jax.Array, jax.Array, float, float]
 jax.config.update("jax_enable_x64", True)
 
 # TODO: Move generically useful math stuff to a new library for reuse
-# TODO: Refactor unit conversion functions
-# TODO: Move units stuff to its own file?
 
 # Constants
 # --------------------------------------------------------
-
 ap = 0.12
 h70 = cosmo.H0.value / 7.00e01
-
-Tcmb = 2.7255
+Tcmb = 2.7255  # K
 kb = const.k_B.value
 me = ((const.m_e * const.c**2).to(u.keV)).value
 h = const.h.value
 Xthom = const.sigma_T.to(u.cm**2).value
-
-Mparsec = u.Mpc.to(u.cm)
-XMpc = Xthom * Mparsec
-
+XMpc = Xthom * u.Mpc.to(u.cm)
 rad_to_arcsec = (180 * 3600) / np.pi
+
 
 # Cosmology
 # --------------------------------------------------------
-dzline = np.linspace(0.00, 5.00, 1000)
-daline = cosmo.angular_diameter_distance(dzline) / u.radian
-nzline = cosmo.critical_density(dzline)
-hzline = cosmo.H(dzline) / cosmo.H0
-
-daline = daline.to(u.Mpc / u.arcsec)
-nzline = nzline.to(u.Msun / u.Mpc**3)
-
-dzline = jnp.array(dzline)
-hzline = jnp.array(hzline.value)
-nzline = jnp.array(nzline.value)
-daline = jnp.array(daline.value)
+zline = jnp.linspace(0, 10, 100)
+daline = jnp.array(
+    (cosmo.angular_diameter_distance(zline) / u.radian).to(u.Mpc / u.arcsec).value
+)
+nzline = jnp.array(cosmo.critical_density(zline).to(u.Msun / u.Mpc**3).value)
+hzline = jnp.array((cosmo.H(zline) / cosmo.H0).value)
 
 
 # Unit conversions
@@ -153,55 +138,59 @@ def y2K_RJ(freq: float, Te: float) -> float:
     return factor * K_CMB2K_RJ(freq)
 
 
-def get_da(z: float) -> float:
+@cache
+def get_da(z: ArrayLike) -> jax.Array:
     """
     Get factor to convert from arcseconds to MPc.
 
     Parameters
     ----------
-    z : float
+    z : ArrayLike
         The redshift at which to compute the factor.
 
     Returns
     -------
-    da : float
+    da : jax.Array
         Conversion factor from arcseconds to MPc.
     """
-    return float(jnp.interp(z, dzline, daline))
+    return jnp.interp(z, zline, daline)
 
 
-def get_nz(z: float) -> float:
+@cache
+def get_nz(z: ArrayLike) -> jax.Array:
     """
     Get the critical density at a given redshift.
 
     Parameters
     ----------
-    z : float
+    z : ArrayLike
         The redshift at which to compute the critical density.
 
     Returns
     -------
-    nz : float
+    nz : jax.Array
         Critical density at the given z.
+        This is in units of solar masses per cubic Mpc.
     """
-    return float(jnp.interp(z, dzline, nzline))
+    return jnp.interp(z, zline, nzline)
 
 
-def get_hz(z: float) -> float:
+@cache
+def get_hz(z: ArrayLike) -> jax.Array:
     """
     Get the dimensionless hubble constant, h, at a given redshift.
 
     Parameters
     ----------
-    z : float
+    z : ArrayLike
         The redshift at which to compute h.
 
     Returns
     -------
-    hz : float
+    hz : jax.Array
         h at the given z.
     """
-    return float(jnp.interp(z, dzline, hzline))
+    return jnp.interp(z, zline, hzline)
 
 
 # FFT Operations
@@ -259,269 +248,6 @@ def tod_hi_pass(tod: jax.Array, N_filt: int) -> jax.Array:
 
 # Model building tools
 # -----------------------------------------------------------
-def make_grid(
-    r_map: float,
-    dx: float,
-    dy: Optional[float] = None,
-    dz: Optional[float] = None,
-    x0: float = 0,
-    y0: float = 0,
-) -> Grid:
-    """
-    Make coordinate grids to build models in.
-    All grids are sparse and are `int(2*r_map / dr)` in each the non-sparse dimension.
-
-    Parameters
-    ----------
-    r_map : float
-        Size of grid radially.
-    dx : float
-        Grid resolution in x, should be in same units as r_map.
-    dy : Optional[float], default: None
-        Grid resolution in y, should be in same units as r_map.
-        If None then dy is set to dx.
-    dz : Optional[float], default: None
-        Grid resolution in z, should be in same units as r_map.
-        If None then dz is set to dx.
-    x0 : float, default: 0
-        Origin of grid in RA, assumed to be in same units as r_map.
-    y0 : float, default: 0
-        Origin of grid in Dec, assumed to be in same units as r_map.
-
-    Returns
-    -------
-    x : jax.Array
-        Grid of x coordinates in same units as r_map.
-        Has shape (`int(2*r_map / dr), 1, 1).
-    y : jax.Array
-        Grid of y coordinates in same units as r_map.
-        Has shape (1, `int(2*r_map / dr), 1).
-    z : jax.Array
-        Grid of z coordinates in same units as r_map.
-        Has shape (1, 1, `int(2*r_map / dr)`).
-    x0 : float
-        Origin of grid in RA, in same units as r_map.
-    y0 : float
-        Origin of grid in Dec, in same units as r_map.
-    """
-    if dy is None:
-        dy = dx
-    if dz is None:
-        dz = dx
-
-    # Make grid with resolution dr and size r_map
-    x = (
-        jnp.linspace(-1 * r_map, r_map, 2 * int(r_map / dx))
-        / jnp.cos(y0 / rad_to_arcsec)
-        + x0
-    )
-    y = jnp.linspace(-1 * r_map, r_map, 2 * int(r_map / dy)) + y0
-    z = jnp.linspace(-1 * r_map, r_map, 2 * int(r_map / dz))
-    x, y, z = jnp.meshgrid(x, y, z, sparse=True, indexing="ij")
-
-    return (x, y, z, x0, y0)
-
-
-# TODO: make this not tied to minkasi, make_grid_from_wcs?
-def make_grid_from_skymap(
-    skymap,
-    z_map: float,
-    dz: float,
-    x0: Optional[float] = None,
-    y0: Optional[float] = None,
-) -> Grid:
-    """
-    Make coordinate grids to build models in from a minkasi skymap.
-    All grids are sparse and match the input map and xy and have size `int(2*z_map/dz)` in z.
-    Unlike `make_grid` here we assume things are radians.
-
-    Parameters
-    ----------
-    skymap : minkasi.maps.Skymap
-        The map to base the grid off of.
-    z_map : float
-        Size of grid along LOS, in radians.
-    dz : float
-        Grid resolution along LOS, in radians.
-    x0 : Optional[float], default: None
-        Map x center in radians.
-        If None, grid center is used.
-    y0 : Optional[float], default: None
-        Map y center in radians. If None, grid center is used.
-
-    Returns
-    -------
-    x : jax.Array
-        Grid of x coordinates in radians.
-        Has shape (`skymap.nx`, 1, 1).
-    y : jax.Array
-        Grid of y coordinates in radians.
-        Has shape (1, `skymap.ny`, 1).
-    z : jax.Array
-        Grid of z coordinates in same units as radians.
-        Has shape (1, 1, `int(2*z_map / dz)`).
-    x0 : float
-        Origin of grid in RA, in radians.
-    y0 : float
-        Origin of grid in Dec, in radians.
-    """
-    # make grid
-    _x = jnp.arange(skymap.nx, dtype=float)
-    _y = jnp.arange(skymap.ny, dtype=float)
-    _z = jnp.linspace(-1 * z_map, z_map, 2 * int(z_map / dz), dtype=float)
-    x, y, z = jnp.meshgrid(_x, _y, _z, sparse=True, indexing="ij")
-
-    # Pad so we don't need to broadcast
-    x_flat = x.ravel()
-    y_flat = y.ravel()
-    len_diff = len(x_flat) - len(y_flat)
-    if len_diff > 0:
-        y_flat = jnp.pad(y_flat, (0, len_diff), "edge")
-    elif len_diff < 0:
-        x_flat = jnp.pad(x_flat, (0, abs(len_diff)), "edge")
-
-    # Convert x and y to ra/dec
-    ra_dec = skymap.wcs.wcs_pix2world(
-        jnp.column_stack((x_flat, y_flat)), 0, ra_dec_order=True
-    )
-    ra_dec = np.deg2rad(ra_dec)
-    ra = ra_dec[:, 0]
-    dec = ra_dec[:, 1]
-
-    # Remove padding
-    if len_diff > 0:
-        dec = dec[: (-1 * len_diff)]
-    elif len_diff < 0:
-        ra = ra[:len_diff]
-
-    if not x0:
-        x0 = (skymap.lims[1] + skymap.lims[0]) / 2
-    if not y0:
-        y0 = (skymap.lims[3] + skymap.lims[2]) / 2
-
-    if x0 is None or y0 is None:
-        raise TypeError("Origin still None")
-
-    ra -= x0
-    dec -= y0
-
-    # Sparse indexing to save mem
-    x = x.at[:, 0, 0].set(ra)
-    y = y.at[0, :, 0].set(dec)
-
-    return x, y, z, float(x0), float(y0)
-
-
-@jax.jit
-def transform_grid(
-    dx: float,
-    dy: float,
-    dz: float,
-    r_1: float,
-    r_2: float,
-    r_3: float,
-    theta: float,
-    xyz: Grid,
-):
-    """
-    Shift, rotate, and apply ellipticity to coordinate grid.
-    Note that the `Grid` type is an alias for `tuple[jax.Array, jax.Array, jax.Array, float, float]`.
-
-    Parameters
-    ----------
-    dx : float
-        Amount to move grid origin in x
-    dy : float
-        Amount to move grid origin in y
-    dz : float
-        Amount to move grid origin in z
-    r_1 : float
-        Amount to scale along x-axis
-    r_2 : float
-        Amount to scale along y-axis
-    r_3 : float
-        Amount to scale along z-axis
-    theta : float
-        Angle to rotate in xy-plane in radians
-    xyz : Grid
-        Coordinte grid to transform
-
-    Returns
-    -------
-    trasnformed : Grid
-        Transformed coordinate grid.
-    """
-    # Get origin
-    x0, y0 = xyz[3], xyz[4]
-    # Shift origin
-    x = (xyz[0] - (x0 + dx / jnp.cos(y0 / rad_to_arcsec))) * jnp.cos(
-        (y0 + dy) / rad_to_arcsec
-    )
-    y = xyz[1] - (y0 + dy)
-    z = xyz[2] - dz
-
-    # Rotate
-    xx = x * jnp.cos(theta) + y * jnp.sin(theta)
-    yy = y * jnp.cos(theta) - x * jnp.sin(theta)
-
-    # Apply ellipticity
-    x = xx / r_1
-    y = yy / r_2
-    z = z / r_3
-
-    return x, y, z, x0 - dx, y0 - dy
-
-
-def tod_to_index(
-    xi: NDArray[np.floating],
-    yi: NDArray[np.floating],
-    x0: float,
-    y0: float,
-    grid: Grid,
-    conv_factor: float = 1.0,
-) -> tuple[jax.Array, jax.Array]:
-    """
-    Convert RA/Dec TODs to index space.
-
-    Parameters
-    ----------
-    xi : NDArray[np.floating]
-        RA TOD, usually in radians
-    yi : NDArray[np.floating]
-        Dec TOD, usually in radians
-    grid : Grid
-        The grid to index on.
-    conv_factor : float, default: 1.
-        Conversion factor to put RA and Dec in same units as the grid.
-
-    Returns
-    -------
-    idx : jax.Array
-        The RA TOD in index space
-    idy : jax.Array
-        The Dec TOD in index space.
-    """
-    x0, y0 = grid[-2:]
-    dx = (xi - x0) * jnp.cos(yi)
-    dy = yi - y0
-
-    dx *= conv_factor
-    dy *= conv_factor
-
-    # Assuming sparse indexing here
-    idx = np.digitize(dx, grid[0].ravel())
-    idy = np.digitize(dy, grid[1].ravel())
-
-    idx = np.rint(idx).astype(int)
-    idy = np.rint(idy).astype(int)
-
-    # Ensure out of bounds for stuff not in grid
-    idx = jnp.where((idx < 0) + (idx >= grid[0].shape[0]), 2 * grid[0].shape[0], idx)
-    idy = jnp.where((idy < 0) + (idy >= grid[1].shape[1]), 2 * grid[1].shape[1], idy)
-
-    return idx, idy
-
-
 @jax.jit
 def bilinear_interp(
     x: jax.Array, y: jax.Array, xp: jax.Array, yp: jax.Array, fp: jax.Array
